+++ pkgs/checks/lib/src/extensions/function.dart
@@ -1,7 +1,10 @@
 // Copyright (c) 2022, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
+import 'dart:async';
 import 'dart:convert';
+
+import 'package:meta/meta.dart' as meta;
 
 import '../../context.dart';
 
@@ -75,4 +78,94 @@ extension FunctionChecks<T> on Subject<T Function()> {
     },
     nestedCondition: that,
   );
+}
+
+extension FunctionPrints on Subject<void Function()> {
+  /// Expects that the function prints text when called.
+  ///
+  /// Intercepts calls to [print] while the function is executed and returns
+  /// a [Subject] to check expectations on the captured printed output.
+  ///
+  /// If the function throws synchronously, this expectation will fail.
+  ///
+  /// {@example /example/function/function/prints.dart}
+  Subject<String> prints([Condition<String>? that]) {
+    return context.nest<String>(
+      () => ['prints'],
+      addPredicate: (predicateNoun) => 'prints $predicateNoun',
+      (actual) {
+        final buffer = StringBuffer();
+        final Object? result;
+        try {
+          result = runZoned(
+            actual,
+            zoneSpecification: ZoneSpecification(
+              print: (_, _, _, line) {
+                buffer.writeln(line);
+              },
+            ),
+          );
+        } catch (e, st) {
+          return Extracted.rejection(
+            actual: ['a function that throws'],
+            which: [
+              ...prefixFirst('threw ', postfixLast(' at:', literal(e))),
+              ...indent(LineSplitter.split(st.toString())),
+            ],
+          );
+        }
+        assert(
+          result is! Future,
+          'Function returned a Future. Provide a `Future<void> Function()` to '
+          'check asynchronous prints.',
+        );
+        return Extracted.value(buffer.toString());
+      },
+      nestedCondition: that,
+    );
+  }
+}
+
+extension AsyncFunctionPrints on Subject<Future<void> Function()> {
+  /// Expects that the asynchronous function prints text when called and
+  /// completed.
+  ///
+  /// Intercepts calls to [print] while the function is executed and waits
+  /// for the returned [Future] to complete before checking expectations
+  /// on the captured printed output.
+  ///
+  /// If the function throws synchronously or returns a [Future] that completes
+  /// with an error, this expectation will fail.
+  ///
+  /// {@example /example/function/async_function/prints.dart}
+  @meta.awaitNotRequired
+  Future<Subject<String>> prints([Condition<String>? printCondition]) {
+    return context.nestAsync<String>(
+      () => ['prints'],
+      addPredicate: (predicateNoun) => 'prints $predicateNoun',
+      (actual) async {
+        final buffer = StringBuffer();
+        try {
+          await runZoned(
+            actual,
+            zoneSpecification: ZoneSpecification(
+              print: (_, _, _, line) {
+                buffer.writeln(line);
+              },
+            ),
+          );
+          return Extracted.value(buffer.toString());
+        } catch (e, st) {
+          return Extracted.rejection(
+            actual: ['a function that throws'],
+            which: [
+              ...prefixFirst('threw ', postfixLast(' at:', literal(e))),
+              ...indent(LineSplitter.split(st.toString())),
+            ],
+          );
+        }
+      },
+      printCondition,
+    );
+  }
 }